Move Headers to separate header.h/c++ file
